Fix overwriting createdDate for old chart versions (#18)
By default, `helm repo index` will update all `created` date fields
in the index to the current date. This results in us losing the
creation dates for old chart versions.

One workaround is to create an index with only the new chart version
and merge this with the existing index. This commit updates
`make publish` to do that.