Fix chart rendering when annotations specified
Commit 70b2777edd3b792c30dd025097ee598b64a45efa caused user-provided
annotations to break the `checksum` StatefulSet annotations. The
`{{ with .Values.annotations }}` instruction changes the scope
under which the checksum is calculated when user annotations are present.

This removes the `{{ with .Values.annotations }}` and instead just
writes out the user-provided annotations directly.

I've also added a user-provided annotation to the `values.yaml` used
in the smoke test to verify the fix.

Fixes #37