feat: support existing persistent volume (#105)
Statefulset `volumeClaimTemplates` create new volumes and `PersistentVolumeClaims` which cannot be mutated. This adds the ability to declare a `persistentVolume` as existing and provide the required PV and PVC data for binding a volume to the statefulset. This is especially useful for restoring backup snapshots during disaster recovery efforts, but could also be used for initial data migrations or other scenarios where users would like to deploy this chart with existing CouchDB data.