fix: convert float64s to integers
If a large integer is specifed in values.yaml, it is parsed by Helm
as a float64. For example:

```
couchdbConfig:
  chttpd:
    timeout: 5184000
```

If this is converted to a string, it produces `5.184e+06` instead of
the expected `5184000`.

Given we never expect floats to be specified as configuration values
to CouchDB, we can workaround this erroneus type conversion by
casting the float64 back to an integer when rendering the ini file.

This then correctly outputs:

```
inifile: |
    [chttpd]
    timeout = 5184000
```

Note this will break any genuine uses of float64 types in the
configuration, though I can't think of any cases where this would be
valid.